Fix marker and before class card styles
