+++ www/src/js/views/today/BeforeLessonCard.scss
@@ -1,0 +1,14 @@
+@import '~styles/utils/modules-entry';
+
+.inBetweenClass {
+  border: 1px solid var(--gray-lighter);
+  box-shadow: 0 0.2rem 0.25rem rgba(#000, 0.2);
+
+  p {
+    margin-bottom: 0.4rem;
+
+    &:last-child {
+      margin-bottom: 0;
+    }
+  }
+}
